reafactor: multiple changes
1. rename `InvoicesHandler` -> `InvoiceServiceHandler` to
match `LightningServiceHandler` naming convention
2. Add `Handlers` module to fetch lightning and invoice
services (Mox possible with this)
