--- lib/lnd_client/invoices_handler.ex
@@ -1,8 +1,0 @@
-defmodule LndClient.InvoicesHandler do
-  alias Invoicesrpc.Invoices.Stub
-  alias Invoicesrpc.AddHoldInvoiceRequest
-
-  def add_hold_invoice(%AddHoldInvoiceRequest{} = request, grpc_channel, macaroon) do
-    Stub.add_hold_invoice(grpc_channel, request, metadata: %{macaroon: macaroon})
-  end
-end
